Split hal and lib pci
