+++ chipsec/modules/tools/vmm/pcie_overlap_fuzz.py
@@ -32,7 +32,7 @@ import time
 import random
 
 from chipsec.module_common import BaseModule, ModuleResult
-from chipsec.hal.pci import print_pci_devices
+from chipsec.lib.pci import print_pci_devices
 
 #################################################################
 # Fuzzing configuration
