+++ chipsec/utilcmd/vmm_cmd.py
@@ -37,7 +37,7 @@ import re
 
 from chipsec.command    import BaseCommand
 from chipsec.hal.vmm    import VMM, get_virtio_devices, VirtIO_Device
-from chipsec.hal.pci    import print_pci_devices
+from chipsec.lib.pci    import print_pci_devices
 from chipsec.exceptions      import VMMRuntimeError
 from argparse           import ArgumentParser
